+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  * added in run-operation for monitior grants
  * added in dynamic table into read specifics
  * updated depends_on_ref and depends_on_source to cater for dbt-core 1.9 changes
+ * get_merge_statement updated to cater for dbt-core 1.9 changes
 
 ## v0.3.5 2024-11-18 - Atomic Unit Test Removed